describe project jrapik
There is a course Training on Solving problems in Java on the learning
platform Stepik (https://stepik.org). This course includes of a set of
Java problems. Solving of this problems should help to reinforce the
knowledges received in the course Java from Zero to Junior Plus on the
same platform. And to receive practical Java experience to be hired as a
Java Junior Programmer.

The text provided is typed by me to solve or to prepare to solve the
issues of the course. But it is not intended to be a precise solution of
the only issues of the course or to be based on the only materials of
the course.

This project can be used in accordance with rules and limitation of
License GPLv3+: GNU GPL version 3 or later
(https://gnu.org/licenses/gpl.html). There is NO WARRANTY, to the
extent permitted by law.